facegausblur.py integrated with main.py WITH BUGS
